Project Description
In this project, I had some data on different ages and whether or not they made insurance purchases. I split the data into training and testing dataset, and then ran a regression on the testing data. To check for accuracy of the model, I considered the model score that checks the accuracy of the predictiongs from the training data against the testing data. For this model, I used logistic regression.